Restructure the project. Modulate the projects. Minor fixes and improvements.
