+++ intro-classification/src/main/java/org/sergedb/oop/intro/models/Individual.java
@@ -1,0 +1,6 @@
+package org.sergedb.oop.intro.models;
+
+import java.util.List;
+
+public record Individual(int id, Boolean isHumanoid, String planet, Integer age, List<String> traits) {
+}
