+++ intro-classification/src/main/java/org/sergedb/oop/intro/models/Universe.java
@@ -1,0 +1,9 @@
+package org.sergedb.oop.intro.models;
+
+public enum Universe {
+    STAR_WARS,
+    MARVEL,
+    HITCHHIKERS,
+    LOTR,
+    UNKNOWN
+}
