+++ intro-classification/src/main/resources/input.json
