+++ pom.xml
@@ -6,14 +6,17 @@
 
     <groupId>org.sergedb.oop</groupId>
     <artifactId>oop-java-projects</artifactId>
-    <version>0.1</version>
+    <version>1.0</version>
+    <packaging>pom</packaging>
+
+    <modules>
+        <module>intro-classification</module>
+        <module>working-with-classes</module>
+        <module>coffee-inheritance</module>
+    </modules>
 
     <dependencies>
-        <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-databind</artifactId>
-            <version>2.18.0</version>
-        </dependency>
+
     </dependencies>
