--- src/main/java/org/sergedb/oop/intermediate_tasks/task_three/TaskThree.java
@@ -1,11 +1,0 @@
-package org.sergedb.oop.intermediate_tasks.task_three;
-
-import org.sergedb.oop.intermediate_tasks.task_one.DisplayManager;
-
-public class TaskThree {
-    public static void main(String[] args) {
-        Assistant assistant = new Assistant("Jora");
-
-        assistant.assist();
-    }
-}
