--- src/main/java/org/sergedb/oop/intermediate_tasks/task_two/FileReader.java
@@ -1,11 +1,0 @@
-package org.sergedb.oop.intermediate_tasks.task_two;
-
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-
-public class FileReader {
-    public static String readFileAsString(String filePath) throws IOException {
-        return new String(Files.readAllBytes(Paths.get(filePath)));
-    }
-}
