--- src/main/java/org/sergedb/oop/papers_please/models/Individual.java
@@ -1,17 +1,0 @@
-package org.sergedb.oop.papers_please.models;
-
-import java.util.List;
-
-public record Individual(int id, Boolean isHumanoid, String planet, Integer age, List<String> traits) {
-
-    @Override
-    public String toString() {
-        return "Individual{" +
-                "id=" + id +
-                ", isHumanoid=" + isHumanoid +
-                ", planet='" + planet + '\'' +
-                ", age=" + age +
-                ", traits=" + traits +
-                '}';
-    }
-}
