--- src/main/java/org/sergedb/oop/papers_please/models/Universe.java
@@ -1,9 +1,0 @@
-package org.sergedb.oop.papers_please.models;
-
-public enum Universe {
-    STAR_WARS,
-    MARVEL,
-    HITCHHIKERS,
-    LOTR,
-    UNKNOWN
-}
