+++ working-with-classes/src/main/java/org/sergedb/oop/classes/task_three/TaskThree.java
@@ -1,0 +1,9 @@
+package org.sergedb.oop.classes.task_three;
+
+public class TaskThree {
+    public static void main(String[] args) {
+        Assistant assistant = new Assistant("Jora");
+
+        assistant.assist();
+    }
+}
