+++ working-with-classes/src/main/resources/input.txt
